+++ R/app-server-fit-model.R
@@ -375,8 +375,16 @@ registerFitModelObservers = function(input, output, session, rv, modelFit, reset
 
     researchQuestion = gsub("\"", "\\\"", researchQuestionRaw, fixed = TRUE)
     attr(m, "wmfm_research_question") = researchQuestion
+    attr(m, "wmfm_research_question_route") = buildResearchQuestionRoute(
+      model = m,
+      researchQuestion = researchQuestionRaw
+    )
     followupQuestion = trimws(input$modelFollowupQuestion %||% rv$modelFollowupQuestion %||% "")
     followupClassification = classifyModelFollowupQuestion(followupQuestion = followupQuestion)
+    followupClassification = attachQuestionRouteToModelFollowupPayload(
+      followupQuestion = followupQuestion,
+      followupPayload = followupClassification
+    )
     followupClassification = enrichFollowupPayloadWithLmPrediction(
       model = m,
       followupPayload = followupClassification